+++ test/app/components/inline_variant_component.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class InlineVariantComponent < ViewComponent::Base
+  def call
+    text_field_tag :name
+  end
+
+  def call_inline_variant
+    text_field_tag :inline_variant
+  end
+end