+++ test/app/components/variant_template_and_inline_variant_template_component.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class VariantTemplateAndInlineVariantTemplateComponent < ViewComponent::Base
+  def call_phone
+    content_tag :div do
+      "Inline Variant Template"
+    end
+  end
+end